==> main/notebook/0_index.ipynb <==
# Scripts to re-create results, figures and tables for "Fair Advice"
#### Kristoffer W.Eriksen, SebastianFest, Ola Kvaløy, Oege Dijk

---

### Notebook index:

Below is an overview of the notebooks provided for the research project ''Fair Advice''. It contains code to recreate all tables and figures presented in the paper. 

1. [Creating dataset](1_make_dataset.ipynb)
2. [Sample characterisitcs](2_balance_and_sample.ipynb)
3. [Treatment effects](3_treatment_effects.ipynb)
3. [Treatment effects w. risk](3_treatment_effects_risk.ipynb)
4. [Beliefs](4_beliefs.ipynb)
