==> main/notebook/1_make_dataset.ipynb <==
#### Import modules

import os  # File system handling
import pandas as pd  # Dataframe handling

#### Set project folders

PROJECT_FOLDER = os.path.abspath(os.path.join(os.getcwd(), os.pardir))
RAW_DATA_FOLDER = os.path.join(PROJECT_FOLDER, 'data', 'raw')
PROCESSED_DATA_FOLDER = os.path.join(PROJECT_FOLDER, 'data', 'processed')
FINAL_DATA_FOLDER = os.path.join(PROJECT_FOLDER, 'data', 'final')

#### Load dataframe

DATA =  os.path.join(
    PROCESSED_DATA_FOLDER, 
    'fair_advice.feather'
)

df = pd.read_feather(DATA)
df.info(verbose=True)

#### Make treatment indicator variables for each main effect

df['Risk'] = df['Treatment'].isin([2,4,6]).astype(int)
df['Base'] = df['Treatment'].isin([1,2]).astype(int)
df['Binding'] = df['Treatment'].isin([3,4]).astype(int)
df['Free'] = df['Treatment'].isin([5,6]).astype(int)

#### Make treatment indicator variables for each treatment axis

df['Risk_type'] = df['Risk'].replace({0:'No Risk', 1:'Risk'})
df.loc[df['Base'] == 1, 'Advice_type'] = 'Base'
df.loc[df['Binding']  == 1, 'Advice_type'] = 'Binding'
df.loc[df['Free']  == 1, 'Advice_type'] = 'Free'

### Make INdicator variables for above median age, education and political orientation 

for cov in ['Age', 'Education', 'Political']:
    df[f'{cov}_m'] = (df[cov] > df[cov].median()).astype(int)

#### Save dataframe

df.info(verbose=True)
path = os.path.join(FINAL_DATA_FOLDER, 'fair_advice_final.feather')
df.to_feather(path)

#### Convert to html

!jupyter nbconvert --output-dir='./docs' --to html 1_make_dataset.ipynb
