==> main/notebook/2_balance_and_sample.ipynb <==
## List of tables:

### 1. [Table 2: Background characteristics of subjects](#balance)

#### Imports libraries

import matplotlib.pyplot as plt  # Plotting library
import os  # File system handling
import pandas as pd  # Dataframe handling
import statsmodels.formula.api as smf  # Regressions analysis
 
from statsmodels.iolib.summary2 import summary_col  # Regression output table

#### Set project directory

PROJECT_FOLDER = os.path.dirname(os.getcwd())
FINAL_DATA_FOLDER = os.path.join(PROJECT_FOLDER, 'data', 'final')
TABLES_FOLDER = os.path.join(PROJECT_FOLDER, 'reports', 'tables')
FIGURES_FOLDER = os.path.join(PROJECT_FOLDER, 'reports', 'figures')

#### Pandas settings 

pd.set_option("display.precision", 3)
pd.set_option("display.expand_frame_repr", False)
pd.set_option("display.max_rows", 25)

#### Set plotting style

plt.style.use('classic')

#### Set plotting properties 

bar_kw = dict(kind='bar', color='0.4', alpha=0.8, lw=0.5, width=0.7)
line_kw = dict(kind='line', lw=1, alpha=1, legend=True)
font_kw = dict(fontsize=11, color='k')
grid_kw = dict(linewidth=1, axis="y", zorder=2, antialiased=True)
x_lab_kw = dict(fontsize=11, labelpad=3)
y_lab_kw = dict(fontsize=11, labelpad=3)
error_kw = dict(elinewidth=2, ecolor='0.15')
legend_kw = dict(frameon=False)
tick_kw = dict(size=5,
               which='both',
               direction='out',
               right=False,
               top=False,
               labelbottom=True)

#### Retrieving dataframe

DATA = os.path.join(
    FINAL_DATA_FOLDER, 
    'fair_advice_final.feather'
)

df = pd.read_feather(DATA)
df.info(verbose=True)

#### Table 1: Treatment table <a id='treatment'></a>

treat_table = pd.crosstab(df['Risk_type'], df['Advice_type'], margins=True)\
    .rename({0: 'No', 1:'Yes'}, axis=0)\
    .rename(lambda x: x.capitalize(), axis=1)\
    .rename_axis('Advice type:', axis=1)\
    .rename_axis('Risk', axis=0)\
    .rename_axis('Advice', axis=1)
    
treat_table.style.to_latex(
    os.path.join(TABLES_FOLDER, 'treat_table.tex'), 
)

display(treat_table)

#### Table 1:  Background characterisitcs of subjects <a id='balance'></a>

sel = (df.Risk == 0)

balance = pd.pivot_table(
    df[sel],
    index='Treatment_name' ,
    aggfunc=['mean', 'sem'],
    values=['Age', 'Female', 'Education', 'Political'],
    margins=True
    )\
    .rename({'mean':'Mean', 'sem':'Se'}, level=0, axis=1)\
    .rename(lambda x: x.capitalize(), axis=1)\
    .rename_axis('Treatment')\
    .swaplevel(axis=1)\
    .sort_index(axis=1, level=0)\
    .round(2)

balance.loc[:, 'N'] = df[sel].groupby('Treatment_name').size()
balance.loc['All', 'N'] = df[sel].groupby('Treatment_name').size().sum()

path = os.path.join(TABLES_FOLDER, 'balance_table.tex')
balance.style.format(precision=2).to_latex(path)

display(balance)

#### Check covariate balance using OLS<a id='balance_ols'></a>

ols_results = [
    smf.ols(formula=f'{variable} ~ C(Treatment_name)', data=df).fit(cov_type='HC1')
    for variable in ['Age', 'Female', 'Education', 'Political']
]

auxiliary = {
    'N': lambda x: f'{x.nobs:.0f}',
    'F': lambda x: f'{x.fvalue:.3f}', 
    'P(>F)': lambda x: f'{x.f_pvalue:.2f}',
}

ols_balance = summary_col(
    ols_results,
    stars=True,
    info_dict=auxiliary,
    float_format='%.3f'
)

label_index = {
    'Intercept':'Constant', 
}

ols_balance = ols_balance.tables[0]\
    .rename(mapper=lambda x: x.replace('C(Treatment_name)[T.', '').replace(']', ''), axis=0)\
    .rename(label_index, axis=0)\
    .rename(lambda x: x.capitalize(), axis=1)\
    .rename_axis('Dependent variable:', axis=1)

path = os.path.join(TABLES_FOLDER, 'ols_balance.tex')
ols_balance.style.format(precision=3).to_latex(path)

display(ols_balance)

#### Convert to html

!jupyter nbconvert --output-dir='./docs' --to html 2_balance_and_sample.ipynb
